# perf: improve proto marshaler